Create searchSuggestions util fn and SearchTextLink component.
