+++ src/shared/components/results/SearchSuggestions.tsx
@@ -1,66 +1,15 @@
-import { Link } from 'react-router-dom';
-
 import AdvancedSearchSuggestion from './AdvancedSearchSuggestion';
 import ExactFieldSuggestion from './ExactFieldSuggestion';
 import TaxonomyLevelsSuggestion from './TaxonomyLevelsSuggestion';
 
 import {
-  parse,
-  stringify,
-} from '../../../query-builder/utils/queryStringProcessor';
+  exactMatchSearchTerms,
+  taxonHierarchySearchTerms,
+} from '../../utils/searchSuggestions';
 
 import { Namespace } from '../../types/namespaces';
-import { Clause } from '../../../query-builder/types/searchTypes';
 
 const simpleQuery = /^[a-zA-Z0-9]+$/;
-
-export const exactMatchSearchTerms = ['gene', 'ec'];
-export const taxonHierarchySearchTerms = ['taxonomy_id', 'organism_id'];
-
-type SearchTextLinkProps = {
-  query: string;
-  text: string;
-};
-
-export const SearchTextLink = ({ query, text }: SearchTextLinkProps) => (
-  <Link
-    // eslint-disable-next-line uniprot-website/use-config-location
-    to={{ search: `query=${query}` }}
-  >
-    {text}
-  </Link>
-);
-
-export const modifyQueryWithSuggestions = (
-  query: string,
-  searchType: 'exact' | 'taxon',
-  searchTerms: string[]
-) => {
-  const parsedQuery = parse(query);
-  let searchValue = '';
-
-  const modifiedClauses: Clause[] = parsedQuery.map((clause) => {
-    if (searchTerms.includes(clause.searchTerm.term)) {
-      const queryBit = clause.queryBits;
-      const modifiedQueryBit: Record<string, string> = {};
-      Object.entries(queryBit).forEach(([k, v]) => {
-        if (searchType === 'exact') {
-          modifiedQueryBit[`${k}_exact`] = v;
-        } else if (searchType === 'taxon') {
-          const suggestion =
-            k === 'taxonomy_id' ? 'organism_id' : 'taxonomy_id';
-          modifiedQueryBit[suggestion] = v;
-        }
-        searchValue = v;
-      });
-      return { ...clause, queryBits: modifiedQueryBit };
-    }
-    return { ...clause };
-  });
-
-  const modifiedQuery = stringify(modifiedClauses);
-  return { modifiedQuery, searchValue };
-};
 
 const hasMatchingQuery = (terms: string[], query: string) =>
   new RegExp(terms.map((w) => `\\b${w}\\b`).join('|'), 'g').test(query);
